Version 1.2.0: feat: Add villager trade recipes; feat: Add cache for stack group outlines to improve efficiency; feat: Add tooltips for creative mode tabs to display the name of the hovered tab; feat: Add icon for the mod; feat: Add English (US) and Simplified Chinese (Mainland China) language files; style: Remove unused code; fix: Stack group outlines appear on the blank areas of the last page.

+++ common/src/main/kotlin/concerrox/emixx/content/stackgroup/EmiGroupStack.kt
@@ -40,7 +40,7 @@ class EmiGroupStack(val group: StackGroup) : EmiStack() {
                 fill(x - 1, y - 1, ENTRY_SIZE, 1, 0xFFFFFFFF.toInt())
                 fill(x + ENTRY_SIZE - 2, y - 1, 1, ENTRY_SIZE, 0xFFFFFFFF.toInt())
                 fill(x - 1, y + ENTRY_SIZE - 2, ENTRY_SIZE, 1, 0xFFFFFFFF.toInt())
-                fill(x, y, ENTRY_SIZE - 2, ENTRY_SIZE - 2, 0x30FFFFFF.toInt())
+                fill(x, y, ENTRY_SIZE - 2, ENTRY_SIZE - 2, 0x30FFFFFF)
             }
 
             if (items.size >= 3) GuiGraphicsUtils.renderItem(raw, items[2].itemStack, x + 3.5F, y + 0.5F, 12F)
